Implement basic aggregation on in-memory virtual tables
This commit does a freakin' lot. We are now basically able to aggregate Kafka messages and materialize the current state in virtual tables in memory.
Of course, in the current state, it's still pretty unstable since there are a lot of assumptions being made and nearly no error handling.

Assumptions include:
- The message key is always a valid UUID
- The message payload is always a valid Avro Value (either Map or Record)
- The payload fields always match some of the Table fields
- There is only one hard-coded table
- The topic is always "appuser"

This list is not exhaustive. I am also using a lot of hacks, mostly because I don't know yet how to do it properly in Rust.